Check error while globbing
Currently when globbing the returned error is simply omitted.
If in parsed directory will be dangling symlink this produces no
error but creates violated build.ninja file.

There is a need to check the error and exit if it appears.

Check the error while globbing and report error of the module.

Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>
Change-Id: I08e813ca23851f8dc07c731e0e12cbedd84371f6