Remove unused properties out of the kernel module
Kernel module consists of `Build` structure which provides many
properties that are not used by the module at all.
Get rid of useless properties and rearrange used properties.

`BuildProps` structure is like a big pot with all properties.
This complicates operating on particular props for different modules
and forces them to contain properties which they do not use at all.
Separate properties which are common for multiple module types and
create `CommonProps` structure for them. Adapt it for every involved
module.

Move `pathMutator` down in the list to be run after `targetMutator`.
This will prevent of unnecessary calls of `processPaths` for every
host:{} and target:{} part. `processPaths` will be called for a module
after all features and defaults will be applied and also modules will
be split to desired variants.

Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>
Change-Id: I7a57a5bfaa6526157c7842e94d4b4bc00a58b0d5