Remove virtualenv
Virtualenv creation is cheap, and virtualenv itself has an issue where
it will not overwrite existing Python instances in a virtualenv's `bin`
directory. This causes env activation to fail, breaking our tests.

In theory we could just clear Travis' caches, but this doesn't seem to
be having any effect. So instead, work around this by removing the
virtualenv directory first.

Change-Id: I8cff58b75c7b2d7637bd002f7f5a6ef7af405ff8
Signed-off-by: Chris Diamand <chris.diamand@arm.com>