Fix macOS rebootstrap issue
When rebootstraping an existing build directory to use a different
working directory, it appears that the microfactory built with the
previous working directory doesn't work on macOS. On detecting the
working directory change, delete the microfactory binary so that it
gets rebuilt.

Testing on macOS is reenabled. This highlighted a failure in
tests/relative_path_tests.sh. '/usr/include' isn't present on macOS,
so just use /bin/bash for testing relative paths to absolute targets.

Change-Id: Ifc25cde20b989ab63514667cac9340fa198323be
Signed-off-by: David Kilroy <david.kilroy@arm.com>