Expand match_srcs feature for cflags, conlyflags and cxxflags
match_srcs was only implemented for ldflags. However, in some
cases it is also convenient to use this features for cflags.
For instance, being able to give a sanitizer file to the
compiler as such:

    cflags: [
        "-fsanitize-blacklist={{match_srcs 'sanitize.blacklist'}}"
    ]

avoids using an environment variable and causes further UX issues.

This commit expands match_srcs feature for cflags, conlyflags
and cxxflags. It also expands the existing match_srcs test.

Signed-off-by: Francois Berder <francois.berder@arm.com>
Change-Id: Idf3960047384b00f066f321bc4b93687a6960485