config_system: Fix integer default handling
When a configuration value is an integer, has a prompt, but hasn't
been set "by the user" we were setting the value as a string when
being read from a file. This breaks handling of defaults (and other
value comparisons).

Make sure that the value is converted to an integer.

Change set_config() to warn about all values that are detected with
the wrong type.

Change-Id: I011155324fca03ddb51da57a9ca197cea7ade524
Signed-off-by: David Kilroy <david.kilroy@arm.com>