Respect build_by_default in kernel modules
Kernel modules on Linux currently hard-code `optional` to `false`,
regardless of the setting of `build_by_default`. This means they are
always built when enabled, even if explicitly requested to be excluded
from the default build.

Fix this by using `!isBuiltByDefault()` when generating the phony rule,
and setting the actual build targets to `optional: true`.

Change-Id: I6d9acd6e4c607463958566da69d5ef20f3c93356
Signed-off-by: Chris Diamand <chris.diamand@arm.com>