Disable CCACHE_BASEDIR to fix depfiles
Bob currently sets CCACHE_BASEDIR to the project root, which in theory
increases cache hits by allowing caches to be used across multiple
copies of the source tree.

Unfortunately, this also causes ccache to rewrite paths inside depfiles
so that they are relative to CCACHE_BASEDIR. As a result, the
now-relative depfile paths do not match the corresponding `build.ninja`
paths, which can be absolute.

As a result, dependencies can be missed on builds using absolute paths,
causing incorrect builds.

Avoid this by explicitly unsetting CCACHE_BASEDIR.

Also remove the warning added in `b0c5003` now that the issue is fixed.

Change-Id: I4d34b60b59a90be42b7d2cbafe65cfb4b42c7574
Signed-off-by: Chris Diamand <chris.diamand@arm.com>