Match lines in the Soong source instead of git SHAs
Handle Android checkouts without complete commit history by matching
specific lines in Soong source files. The fragility of this approach is
significantly reduced by the fact that Go code is always formatted by
`gofmt`, so any changes to the relevant lines are highly likely to be
interface changes that would need accounting for anyway.

Remove the git-SHA-based implementation, because the new approach turns
out to be simpler anyway and will work in the scenarios handling by SHA
detection as well.

Change-Id: I70ba67e037475ac771dfeb57d4cabf388e99fb3c
Signed-off-by: Chris Diamand <chris.diamand@arm.com>