Fix dependency issue caused by bob_resource in Android bp builds
A bob_resource creates a Blueprint module for each element
in its sources. The name of these modules is generated
using the name of the bob_resource and the path of the source.
This is causing somes issues because other modules are refering
to these modules using the bob_resource name instead of the
generated modules names.

The solution consists in modifying ccModuleName which converts
bob module names to Blueprint module names. This function
returns a list of names instead of a single name because a
bob_resource can generate multiple Blueprint modules.
Blueprint modules that depends on a bob_resource now contains
the list of generated module names instead of the bob_resource
name.

Signed-off-by: Francois Berder <francois.berder@arm.com>
Change-Id: I1e62dd3a2cbfc37af7a1a402e5ad9b7cb37d92d2