Fix include and cflag ordering for defaults
cflag and include ordering in defaults is currently inconsistent. This
commit attempts to fix this.

In a hierarchy of defaults, a->b->c, the fields set in leaf nodes (c)
have the lowest priority. Therefore cflags from b must follow cflags
from c (which is currently the case). Include paths from b must come
before include paths from c (this is not currently the case).

In sibling defaults, ["a", "b", "c"], the last element overrides
earlier elements. Therefore cflags from b must follow cflags from a
(this is not currently the case). Include paths from b must come
before include paths from a (which is currently the case).

Address the fact that cflag and include properties have different
ordering requirements by adding field tags to fields read by
Blueprint. The common case for properties is cflag orderring (later
elements override earlier elements), so use the tag
`bob:"first_overrides"` to identify where include orderring is needed.
Note that where libraries are specified through defaults, these also
want include ordering - as the linker is searching for symbols in
libraries in the specified order. For static libraries this will be
overriden if library dependency information is available.

Define functions to select orderring based on the field tag.

Modify the DefaultApplierMutator to separate hierarchical and sibling
propagation.

It's possible that a defaults module occurs more than once within a
defaults hierarchy. When this happens we want to avoid repeating
includes/cflags etc more than once on the command line.

In order to accomplish this, expand the defaults used by each module
into a flat list (maintaining priority order), and tell Blueprint
about this list instead of the full hierarchy.

When expanding properties, we don't need to expand properties on
defaults modules - as the proper modules will always go directly to
each defaults modules to get the required property.

Add a new test to check cflag and include orderring.

Features and Targets are updated to use the same Append functions.

The SupportedVariants mutator is updated to use the same algorithm as
the DefaultApplierMutator. This required updating
defaults.getSplittableProps() to return the right properties instead
of a zero initialised structure.

Change-Id: I396738ede096cad45117cb9a35d794ad237b840c
Signed-off-by: David Kilroy <david.kilroy@arm.com>