Remove "specials" support
Remove support for the `${bob_config}` and `${bob_config_json}`
expansions when used in `srcs`. This is required because Soong does not
accept `srcs` entries that are outside of the source directory, and
therefore this functionality cannot be implemented on `Android.bp`.

The expansions may still be used in `cmd`. However, a depfile must
always be used so that dependencies are tracked correctly. Enforce this
with a new check on all three backends.

Also remove the use of the same expansions in `post_install_cmd`,
because post install commands are also not supported on Soong, and in
any case do not support depfiles.

Change-Id: I9c9f8b77f5881a2c5cc15f6b40a35aa4bbb99486
Signed-off-by: Chris Diamand <chris.diamand@arm.com>