Move `getTarget` into `splittable`
Modules that can be split into target and host-specific variants must
also support a method to discover which variant they are, so move the
corresponding `getTarget` method into `splittable`.

Add a property to external libraries to allow them to remember which
variant they have been set to. The other splittable types (generated
libraries/binaries, and normal libraries/binaries) already support this,
so are unaffected.

Change-Id: Id89edc088386fbde32664ce6a4d5f9a94b4b6795
Signed-off-by: Chris Diamand <chris.diamand@arm.com>