Separate target-specific static lib processing
The graph traversal algorithm for optimal static library ordering
currently uses the same `graph.Graph` instance instance for both target
and host libraries. This means that libraries which should only be
pulled in on one target type may end up incorrectly included on the
other.

This is obviously wrong - in particular on Android, where there are lots
of system libraries on the target which may not be present on the host.

Fix this by using two different `graph.Graph` instances to calculate the
static library ordering.

Add a test case.

Change-Id: Ib5c4b969b168d6a11f595b3c9369dee8d57a34f7
Signed-off-by: Chris Diamand <chris.diamand@arm.com>