Delimit BOB_CONFIG_PLUGINS plugins with a colon ':'
Plugins in a BOB_CONFIG_PLUGINS variable are delimited
with a space but it may happen that plugin names will
also contain the space within it. This will lead to
incorrect tokenisation of the plugin paths.

Change the delimiter character from space to ':' to
avoid possible errors in case plugin name will contain
a space.

Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>
Change-Id: Ida2281f00c0d43e0c265b53bfb3cd569a08a5f3e