Add Warning logger
Provide a warning logger which allow to issue
an alert of some condition where it doesn't
rise an exception and terminate the program.

WarningLogger is able to write to `io.Writer`.
By default Bob will write to `.bob.warnings.csv`
file located in the build directory. Output
file is of CSV format.

There is a BOB_LOG_WARNINGS variable to be
set on bootstrap which allows to set filters
for particular warning categories.
It is a space separated string of format:
"UserWarning:E *:W"

By default all warnings are ignored,
i.e. will be written to log file but not
printed to `io.Stderr`.

Change-Id: Ic28adcfe532248ded6725f3b601ee50cd53b9943
Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>