Config local paths store to Mconfig parser
The current Mconfig parser crawls the workspace recurisvely.

We need a way to easily integrate the Mconfig parser results
with the Bob Gazelle plugin. The best way is to generate
configuration structure as a JSON output.

If we add the current relative path to each configuration item,
we can very simply be able to register the correct configuration
flags for each module and avoid creating excess rules in the
root Bazel file.

Add a python script which outputs the JSON configuration
for desired Mconfig files.

Change-Id: I64a697bc42b18ba08ab4126dee0f807e5dc57b5a
Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>