feat: Parse Bob's module features
This commit registers `BobModule` with parsed
features for every blueprint module.

This way every `BobModule` will constain
features map with all attributes set fot particular
module.

So far there is no any check if the feature is
enabled or not as also no check if we can support
the module.

Change-Id: Id0ed8071a60b28aa569c83000d400634f891e336
Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>