refactor: add source provider interface
This commit adds a clear provider/consumer interface
for sources to be used between different module types.

With this change the source dependancies are resolved
purely via dep tags and do not require the intermediate
flattening of filegroup deps.

It allows any mixing of downstream source providers
with ease.

This commit removes the support for globbing directories,
it reuses the existing `glob` function to simplify the module
with directory support to be added into the `glob` function
at a later time.

Due to changes in how Sources are returned to modules,
the metadata generation has also changed to only list
resolved sources for current module. Should the user
require all transitive sources, the `deps` field can be
used to recursively gather all source files via script.

Change-Id: I49d67bcdef724a2a765945a76fd361f54ebc826a