fix: correct versioned shared library symlinks order
Generated build.ninja from time to time gives different
output for versioned libraries. This comes from the
`symlink` rules which are grabbed from a map which
gives different input of stored data.
Thus the symlink content from a map has to be sorted
before adding Blueprint rules.

The change fix the order for generated ninja symlink
rules to be deterministic.

Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>
Change-Id: Ic1c970ae858fb504d21bf29f3f8f1bf9e3e96602