fix(genrule): fix `tool_files` property
`bob_genrule` module had a broken property set while
using a `ModuleGenerateSource` as a proxy, which is
a temporary solution now.

Additionaly the change sets a dependency when a separate
module is used in `tool_list`, e.g. other `bob_genrule`
used as a tool (with `:` notation).

`bob_genrule` has to be featurable to allow a transition
from `bob_generate_source` modules, as those very higly
are using features for its properties.

Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>
Change-Id: I430092016db179fb8906dc74ff12a6c7d8b98fe7