fix(gendiffer): fixes crashes on whitespace changes
It is possible for the files to differ in raw bytes
but not provide a meaningful diff based on the flags used.
This fix checks if the diff run reports the files to be
the same and returns without error.

In our particular configuration this was causing a crash
when a change resulted in a single space character diff
between files which then resulted in out of bounds access
after spliting the empty diff slice.

Change-Id: Id3dd0b6bb4870e40281bc6fc8787308e3da04822