feat(cclibs): make bob_library splittable
This commit makes `bob_library` behave in the same manner
as legacy libraries when treating `host`/`target` props.

This means that props which rely on selective application
to a given variant will need to be hand migrated in Bazel.

There is no easy way for Bob to disable this feature at
this time.

Change-Id: Iae75c962b9aea6dcc7db60628a102cda43663c32