feat: forward tags from bob_toolchain
Currently only the `owner:` tag is used within Bob,
for this reason this is the only tag currently forwarded
in the backend.

Should a need for wider tag forwarding arise this will
require changes to `Tagable` interface to retrieve the
property from direct deps via the toolchain tag.

Care must be taken as this can result in duplicate
tags, which will produce an error in the case of `owner:`.

Change-Id: I514f06138fd6d5b9ba9fa7402c50af42a42ecf06