On-demand feature struct allocation
Bob's conditionals in modules ("features") can use significant amounts
of RAM during build.bp parsing - each config option spawns a new struct
containing a copy of each build property, resulting in a multiplicative
impact on memory consumption.

Fortunately, close inspection of Blueprint's "unpack" code reveals that
we can instead pass in a nil pointer to the struct, instead of
initialising the properties for each conditional feature, which will
cause property structs to be allocated only when actually used in the
build.bp file.

Use this in Bob by making the property struct for each feature a nil
pointer to the property struct type, instead of a fully-instantiated
interface{} wrapped in a BlueprintEmbed.

Signed-off-by: Chris Diamand <chris.diamand@arm.com>
Change-Id: I42e432fcaa8f1f5cd48ba01c53cb95ce5109aa07