feat(androidNinja): initial toolchain options
This adds some basic flags for the android out
of tree clang toolchain.

'aarch64-linux-android10000' is the abi to
target. I believe 10000 matchs 'current'.
This ABI target is ripped from a bsp build.

'-march=armv8-a+crypto+sha2' target arch taken
from bsp. Further investigations warranted for this.

'-fPIC' this mimics android and is needed for the
linker.

'nostdlibinc' Do not include system sysroot as
we do not provide one.

Change-Id: I9542143c4af46989a514303ef417bf9e30a282a4
Signed-off-by: Brandon Jones <brandon.jones@arm.com>