fix: unset choice
A choice config has always one option to be set.
When the default option for the choice was trying
to unset and no other option was set for that choice
a misleading error was returned saying the value is
wrong and should be one of ['Y', 'y', 'N', 'n'].

Verify the choice config has correct option set
and return proper error message.

Change-Id: I703c94b20e72f9b1b572e4822494e9a3ced645ad
Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>