feat: allow `bazel-bob-build.exe` on `PATH`
We can build `bob` with Bazel.

This prevents having to have the `go` toolchain installed on the system.

Detecting if `bazel-bob-build` is available on `PATH` allows us to avoid
building `bob` and use the one provided.

Using `bazel-bob-build` prevents and accidental conflicts with normal
tools.

This is hacky. Ideally, we'd pass this in via a `bootstrap` argument.

Signed-off-by: Brandon Jones <brandon.jones@arm.com>
Change-Id: I3e0bd1c2643daaa282baad249f02323f1991c09a